Checkpoint commit.  Working internal mixing modes. Receiver input is converted to a fixed point int32_t range [-10,000 ... 10,000] Mixing will be done in this fixed point space to avoid floating pt operations. Setup start of managing and activating mixing modes and gains.

+++ APM2_SensorHead.ino
@@ -107,7 +107,7 @@ bool binary_output = false; // start with ascii output (then switch to binary if
 
 APM_RC_APM2 APM_RC;
 int receiver_raw[NUM_CHANNELS];
-float receiver_norm[NUM_CHANNELS];
+int32_t receiver_norm[NUM_CHANNELS]; // normalized to fixed point [-10000, 10000] range.
 int actuator_pos[NUM_CHANNELS];
 
 #  define CONFIG_MPU6000_CHIP_SELECT_PIN 53
